+++ nway_tree/bench.cpp
@@ -124,15 +124,15 @@ int main(int argc, char* /*argv*/[])
     }
     while( 1 )
     {
-        size_t base = bench< sa::aligned_vector< int32_t >,
+        uint64_t base = bench< sa::aligned_vector< int32_t >,
                              container_only,
                              sa::sse_tag >( "lower_bound .", runSize, loop );
 
-        size_t index1 = bench< sa::aligned_vector< int32_t >,
+        uint64_t index1 = bench< sa::aligned_vector< int32_t >,
                                simd_algorithms::nway_tree::index,
                                sa::sse_tag >( "index SSE ...", runSize, loop );
 
-        size_t index2 = bench< sa::aligned_vector< int32_t >,
+        uint64_t index2 = bench< sa::aligned_vector< int32_t >,
                                simd_algorithms::nway_tree::index,
                                sa::avx_tag >( "index AVX ...", runSize, loop );
